fix(prep): address round-1 review findings on search_prep_notes
- Fix docx chunking: textutil separates paragraphs with a single
  newline (verified against a real Word-style RTF), not the blank-line
  convention PrepMaterialChunker splits on, so a whole .docx extracted
  as one oversized, unsplittable chunk. Normalize each non-empty line
  to its own paragraph before chunking.
- Fix a Process pipe deadlock: docx extraction left standardError as an
  undrained Pipe() while blocking on stdout's readDataToEndOfFile() —
  textutil writing enough stderr output to fill the pipe would hang
  forever. Route stderr to /dev/null, matching the existing
  SyntheticSpeechFixtures.swift precedent for exactly this hazard.
- Track and cancel the index-build Task on Stop, mirroring how history
  compaction is already cancelled — an untracked task kept reading
  files and spawning textutil subprocesses after session teardown, and
  its strong capture of the old CoachDriver kept it alive longer than
  necessary. Also stop dispatching new per-file extraction tasks once
  cancelled (already-dispatched ones still finish; a corpus mid-build
  doesn't retry from scratch).
- Split PrepMaterialChunk.swift and PrepMaterialSearching.swift into
  one type per file, matching AGENTS.md and the Screen/ precedent this
  code's own comments claimed to mirror.
- Carry a successful search_prep_notes result forward via
  work.observations, same as capture_screen, so a failed follow-up
  request in the same attempt doesn't force a fresh retry to re-search
  from scratch.

Verified the docx fix empirically: reproduced the reviewer's exact
scenario (an RTF with ordinary, non-blank-line paragraph breaks
converted to docx) and confirmed textutil's real output has zero
blank-line separators, then confirmed the fix restores correct
per-paragraph splitting.
